requirements.txt
amino_new.py
colored
pyfiglet